Summary: 1. Add some hint for form insert new product. 2. Fix for bug at update commission methods. But it make new bug: if there are some images, the one of them will be stored. 3. Get category for product.

+++ src/test/java/com/google/code/magja/service/category/CategoryRemoteServiceTest.java
@@ -110,7 +110,7 @@ public class CategoryRemoteServiceTest {
 	 */
 	@Test
 	public void testSearch() throws Exception {
-		Category category = service.getTree(new Integer(2));
+		Category category = service.getTree(new Integer(81));
 
 		if (category != null) {
 			if(category.getChildren() != null) {
